fixed unsigned int warning
fixed unsigned int warning by casting to float

+++ src/CreateWindow.cpp
@@ -96,5 +96,5 @@ float rsLib::CreateWindow::getFps()
 	}
 
 	++frame;
-	return fps;
+	return (float)fps; // casting into float because unsigned int warning
  }
